pyproject.toml
[project]
name = "mmk-panel"
version = "0.1.0"
description = "The official Django backend for the MMK online game."
requires-python = ">=3.13"
dependencies = [
 "boto3>=1.42.96",
 "django>=6.0.4",
 "django-storages>=1.14.6",
 "django-stubs>=6.0.3",
 "django-unfold>=0.90.0",
 "djangorestframework>=3.17.1",
 "pillow>=12.2.0",
 "psycopg2-binary>=2.9.12",
 "pyright>=1.1.409",
 "python-dotenv>=1.2.2",
 "ruff>=0.15.12",
]

[build-system]
requires = ["uv_build>=0.11.0,<0.12.0"]
build-backend = "uv_build"
